requirements.txt
python-telegram-bot==20.7
asyncpg==0.29.0
matplotlib==3.8.2
APScheduler==3.10.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
motor==3.7.1
Pillow==10.1.0
pyrogram
tgcrypto
